requirements.txt
Django==1.5.1
psycopg2==2.5
South==0.8.1
djangorestframework==2.3.6
django-haystack==2.0.0
requests==1.2.3
simplejson==2.3.2
pyelasticsearch=0.6
